Delineante/requirements.txt
flask
pillow
requests
huggingface-hub
openai
python-dotenv
